# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the SITCON Camp 2026 game.",
        "title": "Camp 2026 Game API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "User-facing endpoint. Validates the issued game URL token, rotates it, and writes the fresh session token to the camp2026_auth cookie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Authentication"
                ],
                "summary": "User login with auth token",
                "parameters": [
                    {
                        "description": "Auth login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apimodel.AuthLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apimodel.AuthLoginResponse"
                        },
                        "headers": {
                            "Set-Cookie": {
                                "type": "string",
                                "description": "camp2026_auth=\u003cauth-token\u003e; Path=/; HttpOnly; Secure; SameSite=Strict"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    }
                }
            }
        },
...
        "/shop/purchases": {
            "post": {
                "security": [
                    {
                        "AuthCookieAuth": []
                    }
                ],
                "description": "Purchases one enabled shop item, deducts open power, and adds it to the current player's item bag.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shop"
                ],
                "summary": "Purchase shop item",
                "parameters": [
                    {
                        "description": "Purchase request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/shop.PurchaseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/shop.PurchaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apimodel.AuthLoginRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string",
                    "maxLength": 512,
                    "example": "token-1"
                }
            }
        },
        "apimodel.AuthLoginResponse": {
            "type": "object",
            "properties": {
                "player": {
                    "$ref": "#/definitions/apimodel.AuthPlayerSummary"
                }
            }
        },
        "apimodel.AuthPlayerSummary": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string",
                    "example": "https://example.test/avatar/alice.png"
                },
                "nickname": {
                    "type": "string",
                    "example": "Alice"
                },
                "openPower": {
                    "type": "integer",
                    "example": 1280
                },
                "playerId": {
                    "type": "string",
                    "example": "7H9K2Q"
                },
                "role": {
                    "type": "string",
                    "example": "staff"
                },
                "team": {
                    "$ref": "#/definitions/apimodel.AuthTeamSummary"
                }
            }
        },
        "apimodel.AuthTeamSummary": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Blue Team"
                },
                "teamId": {
                    "type": "string",
                    "example": "8M4RXP"
                }
            }
        },
...
        "httpx.ErrorDetail": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httpx.ProblemDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpx.ErrorDetail"
                    }
                },
                "instance": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "shop.PurchaseRequest": {
            "type": "object",
            "required": [
                "itemId"
            ],
            "properties": {
                "itemId": {
                    "type": "string",
                    "example": "item_adventure_backpack"
                }
            }
        },
        "shop.PurchaseResponse": {
            "type": "object",
            "properties": {
                "item": {
                    "$ref": "#/definitions/shop.ShopItemResponse"
                },
                "itemId": {
                    "type": "string",
                    "example": "item_adventure_backpack"
                },
                "openPower": {
                    "type": "integer",
                    "example": 1230
                },
                "priceOpenPower": {
                    "type": "integer",
                    "example": 50
                },
                "purchaseId": {
                    "type": "string",
                    "example": "purchase_abc123"
                },
                "quantity": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "shop.ShopItemResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "冒險背包，可用於小石合成。"
                },
                "iconPath": {
                    "type": "string",
                    "example": "/game-icons/items/item_adventure_backpack.png"
                },
                "id": {
                    "type": "string",
                    "example": "item_adventure_backpack"
                },
                "name": {
                    "type": "string",
                    "example": "冒險背包"
                },
                "priceOpenPower": {
                    "type": "integer",
                    "example": 50
                },
                "rarity": {
                    "type": "string",
                    "example": "common"
                },
                "redeemed": {
                    "type": "boolean",
                    "example": false
                },
                "source": {
                    "type": "string",
                    "example": "shop"
                },
                "type": {
                    "type": "string",
                    "example": "material"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AuthCookieAuth": {
            "description": "User auth cookie. Send as `Cookie: camp2026_auth=\u003cbase64url-auth-token\u003e`.",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}